added cogs, organization yay finally :)
